% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_shrub.R
\name{est_shrub_density}
\alias{est_shrub_density}
\title{Calculate the density of a shrub}
\usage{
est_shrub_density(height, length, width)
}
\arguments{
\item{height}{height of the object}

\item{length}{length of the object}

\item{width}{width of the object}
}
\value{
density The density of the object (numeric)
}
\description{
use the length width and height of an object to calculate its density
returns density as a floating point number
}
