% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManovaModel.R
\name{manova_project}
\alias{manova_project}
\title{Determines the relationship between the hindfoot_length and weight variables by sex concurrently}
\usage{
manova_project(surveys, weight, sex, hindfoot_length)
}
\arguments{
\item{weight}{Weight of rodents}

\item{sex}{Sex of rodents}

\item{hindfoot_length}{Hindfoot length of rodents}
}
\value{
returns hindfoot length and weight as statistical values that determine what effect the independent variable has on the depenedent variable
}
\description{
Use hindfoot length and weight column from surveys dataset
Returns hindfoot length and weight as statistical values together from rodents surveys dataset
}
