% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCount.R
\docType{data}
\name{plot_count}
\alias{plot_count}
\title{Calculate the number of rodents per plot type excluding nas}
\format{
An object of class \code{data.frame} with 5 rows and 2 columns.
}
\usage{
plot_count
}
\arguments{
\item{plot_type}{Indicates which plot used by rodents}
}
\value{
number of rodents found at each plot type
}
\description{
use plot type column from dataset
returns number of rodents present at each plot
}
\keyword{datasets}
