% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCount.R
\name{plot_count_project}
\alias{plot_count_project}
\title{Calculate the number of rodents per plot type excluding nas}
\usage{
plot_count_project(surveys, plot_type)
}
\arguments{
\item{plot_type}{Indicates which plot used by rodents}
}
\value{
number of rodents found at each plot type
}
\description{
use plot type column from dataset
returns number of rodents present at each plot
}
