% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardDev.R
\name{stddev_project}
\alias{stddev_project}
\title{Calculates and summarizes the standard deviation and mean of weight by species id}
\usage{
stddev_project(surveys, species_id, weight)
}
\arguments{
\item{species_id}{Identification of rodents by species type}

\item{weight}{Weight of rodents}
}
\value{
returns the summary of the weight column through standard deviation and mean by species id
}
\description{
Use weight column from surveys dataset and group by species id
Returns hindfoot length and weight as statistical values together from rodents surveys dataset
}
