% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{graphplot_project}
\alias{graphplot_project}
\title{Graph relationship between sex and hindfoot length}
\usage{
graphplot_project(data, sex, hindfoot_length)
}
\arguments{
\item{sex}{Sex of rodents}

\item{hindfoot_length}{Hindfoot length of rodents}
}
\value{
Returns gpp as graph plot between sex and hindfoot length
}
\description{
use sex and hindfoot length from surveys dataset to determine relationship
Returns graphplot for sex versus hindfoot length
}
